Implement a job.input for GenericMaster
